Internal change
PiperOrigin-RevId: 356158181
