+++ official/modeling/progressive/utils.py
@@ -14,6 +14,9 @@
 
 """Util classes and functions."""
 
+from absl import logging
+import tensorflow as tf
+
 # pylint: disable=g-direct-tensorflow-import
 from tensorflow.python.training.tracking import tracking
 
@@ -29,3 +32,25 @@ class VolatileTrackable(tracking.AutoTrackable):
     for k, v in kwargs.items():
       delattr(self, k)  # untrack this object
       setattr(self, k, v)  # track the new object
+
+
+class CheckpointWithHooks(tf.train.Checkpoint):
+  """Same as tf.train.Checkpoint but supports hooks.
+
+  In progressive training, use this class instead of tf.train.Checkpoint.
+
+  Since the network architecture changes during progressive training, we need to
+  prepare something (like switch to the correct architecture) before loading the
+  checkpoint. This class supports a hook that will be executed before checkpoint
+  loading.
+  """
+
+  def __init__(self, before_load_hook, **kwargs):
+    self._before_load_hook = before_load_hook
+    super(CheckpointWithHooks, self).__init__(**kwargs)
+
+  # override
+  def read(self, save_path, options=None):
+    self._before_load_hook(save_path)
+    logging.info('Ran before_load_hook.')
+    super(CheckpointWithHooks, self).read(save_path=save_path, options=options)
